requirements.txt
numpy
nes_py==8.1.1 
gym-super-mario-bros==7.3
wandb
line_profiler
pytorch
